Refactor execution infrastructure so test instance can be set
WARNING: this refactoring makes the testInstance property of
DescriptorBasedTestExecutionContext mutable and also introduces
reflective hacks for setting the testInstance in various places.
This is a sign that we have mixed the responsibilities of the internal
execution infrastructure and the public API for how extensions determine
their context (i.e., the original intent of TestExecutionContext).