Mask checked exceptions as unchecked exceptions in ReflectionUtils
Prior to this commit, ReflectionUtils.handleException() wrapped checked
exceptions in a TargetExceptionWrapper (a subclass of RuntimeException).
This forced clients of ReflectionUtils to manually handle unwrapping of
checked exceptions.

This commit introduces a new throwAsRuntimeException() method in
ReflectionUtils which uses a reflective hack (based on generics and type
erasure) that tricks the Java compiler into believing that the thrown
exception is an unchecked exception. This approach allows both checked
and unchecked exceptions to transparently propagate up the call stack
without forcing callers to explicitly handle checked exceptions. In
addition, this approach avoids the need to wrap/unwrap checked
exceptions via the TargetExceptionWrapper.

Furthermore, this commit refactors exception handling code within the
JUnit 5 engine to use the newly introduced ThrowableCollector API for
consistent, simplified tracking of exceptions.

Issue: #64

------------------------------------------------------------------------
On behalf of the community, the JUnit Lambda Team thanks
Samuraism, Inc. (http://samuraism.com/) for supporting the
JUnit crowdfunding campaign!
------------------------------------------------------------------------