Do not wrap checked exceptions in Assertions.assertAll()
Prior to this commit, assertAll() wrapped a checked exception in a
RuntimeException and threw the RuntimeException. This behavior, however,
is undesirable since it hides the actual cause of the failure.

This commit addresses this issue by 'masking' checked exceptions as
unchecked exceptions without actually wrapping them.

Issue: #91

------------------------------------------------------------------------
On behalf of the community, the JUnit Lambda Team thanks msg systems ag
(http://www.msg-systems.com) for supporting the JUnit crowdfunding
campaign!
------------------------------------------------------------------------