Substitute JUnit version dynamically in the User Guide
With this commit, the {junit-version} is now available in AsciiDoc files
for dynamic substitution of the current version of JUnit (configured in
gradle.properties). The generated User Guide also now includes the
version in the header and footer.

------------------------------------------------------------------------
On behalf of the community, the JUnit Lambda Team thanks Klarna AB
(http://www.klarna.com) for supporting the JUnit crowdfunding campaign!
------------------------------------------------------------------------