Return original object in Preconditions
The commit updates the Preconditions class so that its methods return
the original object in case of a successful check. In addition, various
clients of the Preconditions API have been updated to make use of this
new feature.

Issue: #95

------------------------------------------------------------------------
On behalf of the community, the JUnit Lambda Team thanks Klarna AB
(http://www.klarna.com) for supporting the JUnit crowdfunding campaign!
------------------------------------------------------------------------