Rename FirstTestCase to FirstJUnit5Tests
This will allow FirstJUnit5Tests to be executed with the build once we
switch to the JUnit 5 plugin. In addition, this commit demonstrates how
to include multiple, non-overlapping sections from source code.

------------------------------------------------------------------------
On behalf of the community, the JUnit Lambda Team thanks Klarna AB
(http://www.klarna.com) for supporting the JUnit crowdfunding campaign!
------------------------------------------------------------------------