Introduce source in RegisteredExtensionPoint
In order to improve error reporting and logging, a new 'source'
property has been introduced in RegisteredExtensionPoint.

If an extension point is registered declaratively via @ExtendWith, the
'extensionPoint' and 'source' properties will contain a reference to
the same object; however, if an extension point is registered
programmatically -- for example, by an ExtensionRegistrar as a lambda
expression or by the framework via the TestExtensionRegistry -- the
'source' may be the ExtensionRegistrar that registered the extension
point, the underlying method that implements the extension point API,
or similar.

The source is currently used for error reporting in
ExtensionPointSorter when conflicting extensions are encountered. In
addition, the source is logged under the
'org.junit.gen5.engine.junit5.execution' category at the TRACE log
level whenever an extension is registered.

Issue: #112