Support registration of extensions that implement multiple APIs
Prior to this commit it was (seemingly) not possible to
programmatically register an extension that implemented multiple
ExtensionPoint APIs via the ExtensionPointRegistry with a single
'register(...)' invocation. It was in fact possible, but the user had
to declare that the extension only implemented one of the many APIs
which could lead the user to believe that the extension would not be
registered for all implemented APIs.

This commit remedies this situation by introducing additional
register() methods in ExtensionPointRegistry which make it clear that
an extension can be registered for multiple APIs simultaneously.

Furthermore, the example TimingExtension has been updated to make use of
this new feature.

Issue: #122