Upgrade to Spotless 1.3.3 with Eclipse 4.5.2 support
The upgrade to Spotless 1.3.3 brings with it all of the bug fixes
applied to the Eclipse formatter since Eclipse 4.5 M6.

For example, whitespace is now properly inserted after array type
declarations in annotation attributes as well as before "implements"
in class declarations.

Anyone developing JUnit 5 in Eclipse should therefore upgrade to
Eclipse 4.5.2+ or 4.6+.

Issue: #183