Simplify interface of the visitor.
Every visitor gets the instance of the TestDescriptor and a Runnable for
deleting it from the hierarchy. Everybody who has access to a
TestDescriptor can therefore delete it. That's why I added the delete
method to the TestDescriptor interface because I think that this is
easier to understand compared to the Runnable.