Ensure that repeatable annotations are discovered in superclasses
Prior to this commit, it was impossible to declare `@ExtendWith` on a
superclass and have it discovered. On the contrary, the declaration
would simply be silently ignored.

The same holds true for any other repeatable annotation that is looked
up via JUnit 5's AnnotationUtils.

This commit fixes this problem by completely overhauling the
findRepeatableAnnotations() search algorithm in AnnotationUtils.

Issue: #207