Find repeatable annotations on interfaces
Prior to this commit, the findRepeatableAnnotations() method in
AnnotationUtils did not find repeatable annotations declared on
interfaces (e.g., so called "test interfaces"). As a consequence, test
interfaces could declare default methods but no default configuration
in the form of class-level tags or registered extensions required by
the default methods.

This commit addresses this issue by refining the search algorithm for
findRepeatableAnnotations() to search on interfaces as well.

Issue: #226