Add `throws JUnitException` to method signature for Checkstyle
If a method declares that it throws a particular exception via the
Javadoc `@throws` tag, Checkstyle will fail the build if the method
does not formally declare the exception in the method signature as
well.

This commit therefore updates UniqueId.parse() to declare that it
throws JUnitExeption both via Javadoc and the method signature.

Issue: #251