Apply InstancePostProcessor extensions for every test instance
Prior to this commit, InstancePostProcessor extensions were only applied
during the execution of a MethodTestDescriptor instead of at the point
when such an instance was physically instantiated. Consequently,
InstancePostProcessor extensions were never applied to instances of
outer classes when test methods in a nested class were executed, which
lead to improperly initialized outer instances (e.g,
NullPointerExceptions, etc.).

This commit fixes this bug by relocating the invocation of
InstancePostProcessor extensions to places in the framework where test
instances are physically instantiated (i.e., in ClassTestDescriptor and
NestedClassTestDescriptor).

Furthermore, the argument list for the postProcessTestInstance() method
in the InstancePostProcessor API has been modified to accept the test
instance and a generic ExtensionContext.

Issue: #252