Introduce deactivation mechanism for conditions
Prior to this commit, there was no way to deactivate or turn off
support for conditional test execution. In other words, there was no
way to execute a @Disabled test and no way to deactivate custom
implementations of ContainerExecutionCondition or
TestExecutionCondition.

This commit addresses this issue by introducing support for supplying a
`junit.conditions.deactivate` pattern to the Launcher via a
configuration parameter in the TestDiscoveryRequest.

If the `junit.conditions.deactivate` pattern consists solely of an
asterisk (*), all conditions will be deactivated. Otherwise, the
pattern will be used to match against the fully qualified class name
(FQCN) of each registered condition. Any dot (.) in the pattern will
match against a dot (.) or a dollar sign ($) in the FQCN. Any asterisk
(*) will match against one or more characters in the FQCN. All other
characters in the pattern will be matched one-to-one against the FQCN.

Examples:

- *: deactivates all conditions.
- org.junit.*: deactivates every condition under the org.junit base
  package and any of its subpackages.
- *.MyCondition: deactivates every condition whose simple class name is
  exactly MyCondition.
- *System*: deactivates every condition whose simple class name
  contains System.
- org.example.MyCondition: deactivates the condition whose FQCN is
  exactly org.example.MyCondition.

Issue: #78