Ensure JAR manifests are only generated when necessary
Prior to this commit, JAR manifests were generated with dynamic values
for the buildDate and buildTime for every Gradle run. This resulted in
project JARs being rebuilt even if nothing was changed in the projects.
Since project JARs serve as inputs into the testing tasks of other
projects, the net effect was that tests would always be executed even
if nothing had actually changed.

This commit addresses this issue by modifying the Gradle build so that
JAR manifests are only generated/regenerated if the code for the
corresponding project is compiled/recompiled.

Fixes: #266