Require floating point deltas to be positive
This commit adds a check that the delta argument supplied to
`assertEquals(float, float, float)` and `assertEquals(double, double,
double)` is positive. Same check is applied to methods that accept a
message or message supplier.

This check is needed because it does not make sense to compare two
floating point numbers within negative or zero delta.

Issue: #269