Ensure JUnit Gradle plugin does not influence project classpaths
Prior to this commit, the JUnit Gradle plugin modified the test runtime
classpath of the containing project by adding the junit-console module
(and all of its transitive dependencies) to it. Consequently, the
`junit4` and `junit5` engines were always on the classpath, both for
the project and for the task created by the plugin. The negative side
effect is that those engines were then included in the test plan by
default, even if the user didn't want either of those engines.

This commit addresses this issue by introducing a `junitPlatform`
configuration that serves as a classpath container for the JUnit
Platform modules required by the custom task created by the JUnit
Gradle plugin. This junitPlatform configuration has a dependency on the
junit-launcher (and its transitive dependencies such as the
junit-engine-api) and a dependency on the junit-console (but without
any of its transitive dependencies).

For example, executing `gradlew :junit-tests:dependencies` will result
in a dependency tree similar to the following.

junitPlatform
+--- org.junit:junit-launcher:5.+ -> 5.0.0-SNAPSHOT
|    \--- org.junit:junit-engine-api:5.0.0-SNAPSHOT
|         +--- org.junit:junit-commons:5.0.0-SNAPSHOT
|         \--- org.opentest4j:opentest4j:1.0.0-M1
\--- org.junit:junit-console:5.+ -> 5.0.0-SNAPSHOT

The classpath for the custom task created by the JUnit Gradle plugin is
then composed of the junitPlatform configuration and anything the user
adds to the test run-time classpath for the project.

Consequently:

1) No TestEngines are accidentally added to the test runtime classpath
   of the project.
2) No TestEngines are accidentally added to the classpath of the custom
   task created by the JUnit Gradle plugin.

Closes: #285