Ensure afters are invoked if exception is thrown by test or a before
before: @BeforeEach method, BeforeEachCallback, BeforeTestExecutionCallback

after:  @AfterEach method, AfterEachCallback, AfterTestExecutionCallback

Prior to this commit, if an exception was thrown by a test method or a
"before", then "afters" were never invoked, which is in strict contrast
to the semantics of JUnit 4 and other frameworks.

This commit fixes this problem by ensuring that "afters" (for the
appropriate nesting level) are invoked even if the test method or a
"before" throws an exception.

Fixes: #357