Ensure AfterAlls are invoked if exception is thrown by a BeforeAll
- BeforeAll: @BeforeAll method or BeforeAllCallback
- AfterAll:  @AfterAll method or AfterAllCallback

Prior to this commit, if an exception was thrown by a BeforeAll, then
AfterAlls were never invoked, which is in strict contrast to the
semantics of JUnit 4 and other frameworks.

This commit fixes this problem by ensuring that AfterAlls (for the
appropriate nesting level) are invoked even if a BeforeAll throws an
exception.

Fixes: #359