Make path to Javadoc in User Guide configurable
Prior to this commit, the path to Javadoc on the server was hard
coded to "current".

This commit makes the path configurable as follows.

- Introduced `docsVersion` project property which defaults to "snapshot"
- `docsVersion` is passed to AsciiDoctor as `docs-version`
- `javadoc-root` in user guide is now constructed using `docs-version`

To change the Javadoc path, simply supply a custom value to the Gradle
build. For example, to set the path to 5.0.0-M1, run gradle as follows.

$> gradlew asciiDoc -PdocsVersion=5.0.0-M1

Issue: #319