Make path to source branch/tag in User Guide configurable
Prior to this commit, the path to the branch or tag for source code on
GitHub was hard coded to `master`.

This commit makes the path configurable as follows.

- Introduced `releaseBranch` project property which defaults to `master`
- `releaseBranch` is passed to AsciiDoctor as `release-branch`
- Renamed `master-branch` in user guide to `current-branch`
- `current-branch` in user guide is now constructed using `release-branch`

To change the branch/tag path, simply supply a custom value to the Gradle
build. For example, to set the path to 5.0.0-M1, run gradle as follows.

$> gradlew asciiDoc -PreleaseBranch=5.0.0-M1

Issue: #319