Use real names and class names in generated XML report
Prior to this commit, display names were used as the `name` and
`classname` in XML reports; however, this is not very useful for
tooling. Furthermore, since commit ab20137, display names are now
always included in the `system-out` element.

Consequently, this commit now uses real names and class names in
the generated XML report.

Issue: #372