Allow TestEngines to be completely dynamic
Prior to this commit, it was impossible to create a custom TestEngine
that was completely dynamic in nature (i.e., that did not discover any
tests during the discovery phase) since such an engine would get
pruned from the test plan.

This commit addresses this issue by ensuring that engines without
statically registered children are not pruned by the Launcher.

Closes: #367