Use real names and class names in JUnitPlatform runner
Prior to this commit, display names were used as the `name` and
`className` in Descriptions created by the JUnitPlatform runner. This
behavior had a coolness factor in that emoji and special characters
would be displayed in IDEs and test reports. More importantly, however,
it had a huge negative side effect in that such reports were not
hierarchically structured. Specifically, all test classes appeared to
be in the "default" package, which is not very useful for tooling or
analysis.

Consequently, this commit now uses real method names and class names in
Descriptions for "tests" created by the JUnitPlatform runner if the
source of the test is a Java method. This commit does not alter the
existing support for "suites" (i.e., test containers) or test sources
other than Java methods; in those cases, the display name is still used.

Issue: #372