Introduce @UseTechnicalNames for the JUnitPlatform runner
Due to recent changes, technical names were previously used for test
methods and their test classes instead of display names.

This commit introduces a new @UseTechnicalNames annotation that allows
one to opt-in for the usage of technical names instead of display
names. If @UseTechnicalNames not directly present on the test class run
via the JUnitPlatform runner, display names will be used by default.

Closes #379