Ensure extensions cannot be registered as duplicates
Commit 6c146c5a76 introduced a regression that allowed an extension
implementation to be registered multiple times if registered in both
a parent and child extension context (e.g. at the class-level and
method-level).

This commit fixes this by reinstating logic for querying the parent
extension context as well when adding an extension to a child context.

This commit also improves the documentation on extension inheritance in
the User Guide.

Fixes: #390