Ensure failure in container fails build with JUnit Gradle plugin
Prior to this commit, if a container failed -- for example, if a test
class threw an exception in an @BeforeAll method -- the build would
still pass when using the JUnit Platform Gradle plugin. This was also
true for the ConsoleLauncher, since the Gradle plugin uses the
ConsoleLauncher under the hood.

This commit fixes this bug by ensuring that the TestExecutionSummary
now properly tracks failures in containers as well as in tests. In
addition, TestExecutionSummary supports a new getTotalFailureCount()
method that the ConsoleLauncher uses to determine its exit code.

Fixes: #402