Support primitive types & arrays in ReflectionUtils.loadClass()
Prior to this commit, ReflectionUtils did not support loading classes
by name for primitive types or arrays.

This commit addresses this issue by introducing a static look-up map
from names of primitive types and arrays of primitive types to their
corresponding classes. In addition, explicit support has been added for
loading classes for objects arrays given a name in the form of
`[L<type>;` (e.g., `[Ljava.lang.String;` for a String[]).

More importantly, this commit is a prerequisite for selecting methods
based on _fully qualified method names_ and loading methods that accept
parameters that are primitives or arrays.

Issue: #332