Output exception stack trace in ColoredPrintingTestListener
Prior to this commit, the ColoredPrintingTestListener neglected to
output the actual exception type when printing an exception message,
thereby making it very difficult to debug the underlying cause of the
failure.

This commit ensures that the actual exception type and its stack trace
are output by ColoredPrintingTestListener when printing exception
messages.

Fixes: #460