Reduce configuration time impact of versioning plugin
The call to `versioning.info.commit` in the `allprojects` block
lead to calling `git status` for every project. I replaced this
with a single call that is cached in a property.

This reduces the time to call `gradle help` from 1.6s to 0.7s.

We should consider only including this information on CI builds,
as local builds are unlikely to care about an accurate build revision
in the manifest.