Do not run DeGraph as part of standard test task
Prior to this commit, the slow DeGraph dependency tests were always
executed via the standard Gradle `test` task. The only way to execute
all tests without DeGraph was to execute `gradlew test -x :test` which
is not very intuitive for most developers.

This commit addresses this issue by introducing a new (empty) `degraph`
task in the root project which acts as an "alias" for the
`junitPlatformTest` task.

In addition, the dependency from the standard `test` task to the
`junitPlatformTest` task has been removed for the root project, and a
dependency from the `check` task to the `degraph` task has been added
for the root project.

Consequently, executing `gradlew test` will no longer run the DeGraph
tests, but executing `gradlew degraph` or `gradlew check` will.

Issue: #481