Support checked exceptions in DynamicTest.stream
Prior to this commit, `DynamicTest.stream()` accepted a conventional
`Consumer` for its _test executor_ which made it impossible to supply
a consumer that threw checked exceptions, which is in direct contrast to
the standard support for dynamic tests provided by the `Executable`
interface accepted by `DynamicTest.dynamicTest()`.

This commit fixes this problem by introducing a new `ThrowingConsumer`
that is capable of throwing any kind of exception, and
`DynamicTest.stream()` now accepts accepts a `ThrowingConsumer` instead
of a conventional `Consumer` for its _test executor_, thereby allowing
for custom streams of dynamic tests that may potentially throw checked
exceptions.

Fixes: #486