Ensure @AfterEach methods execute bottom-up w/in class hierarchy
Prior to this commit, a bug had been introduced that caused @AfterEach
methods within a test class hierarchy to be executed top-down instead
of bottom-up.

The cause of this bug was related to the introduction of synthesized
extensions for executing @AfterEach methods transparently as
extensions. Since extensions are registered in-order, the order of
"after" extensions is always reversed before executing them, and this
mechanism reversed the bottom-up ordering of @AfterEach methods.

This commit fixes this bug by reversing the order of @AfterEach methods
before synthesizing them as extensions.

Fixes: #495