Unwrap element when generating reason for @Disabled test
Prior to this commit, the Optional<AnnotatedElement> was converted
directly to a String when generating the reason for a @Disabled test,
which resulted in output similar to the following.

Optional[void example.DisabledTestsDemo.testWillBeSkipped()] is @Disabled

This commit fixes this problem by unwrapping the Optional when
generating the reason, resulting in output similar to the following.

void example.DisabledTestsDemo.testWillBeSkipped() is @Disabled

Fixes: #514