Deprecate expectThrows in favor of assertThrows returning the exception
Prior to this commit, expectThrows() returned the thrown exception;
whereas, assertThrows() simply swallowed the exception. This lead to
confusion for users since it was not readily clear which variant should
be used. In addition, expectThrows() was the only method in the
Assertions class (other than the necessary fail() variants) that was not
named assert*().

This commit addresses this issue by refactoring assertThrows() to return
the thrown exception and deprecating the existing (and now duplicated)
expectThrows() method.

Closes: #531