Honor the build dir in Gradle plugin (#538)
Before this change, there was an assumption that
all users use the project directory as the buildDir.
In some cases, they are re-mapped to a single
buildDir at the root of the project to make management
easier. This change uses the builDir from the project
to set where the output will be.

Also to make the plugin nicer to use, users can set
the reportsDir to be a GString (or any object really)
and rely on the behavior of Project.file(..) to
resolve it to the correct path.