Discover @BeforeAll/@AfterAll on static interface methods
Prior to this commit, ReflectionUtils.getInterfaceMethods() only
detected default methods in interfaces (i.e., by supplying
Method::isDefault as a filter predicate); however, this limiting
behavior was unintentional and resulted in static methods not being
found. For example, @BeforeAll/@AfterAll were simply not supported when
declared on static methods in test interfaces.

This commit addresses this issue by modifying
ReflectionUtils.getInterfaceMethods() so that it returns all methods in
interfaces which are concrete (i.e., not abstract).

Closes: #571