Use canonical names for exception types in assertions
Prior to this commit, if an unexpected exception was thrown from an
Executable passed to Assertions.assertThrows() the standard class names
of the expected and actual exception types were used when generating
failure messages. However, this could lead to ambiguity if the same
exception type was loaded via different class loaders. Similarly, the
canonical name for static nested classes is more readable than the
standard name.

This commit attempts to use canonical names for exception types when
generating failure messages for assertThrows whenever possible. If the
canonical name of an exception type is not retrievable, the standard
class name will be used as a fall back.

Issue: #575