Use parent display name in surefire reporting if class is unknown.
If no test source was attached to test identifier, the adapter
used its unique Id to report as source string to surefire.

This unique id is for internal use though, and can be a base64 encoding.

Replaced with the more meaningful displayname of the _parent_.
To facilitate obtaining the parent, the test plan is used.

Also added a test case that explicitly checks the name that is reported
to surefire as source.