Change ConsoleLauncher execution result from int to summary
Before this commit, the result of ConsoleLauncher#execute(...) was
reduced to an int primitive, the exit code used by ConsoleLauncher#main.
That reduction limited the use of the ConsoleLauncher for integration
tests, as the TestExecutionSummary created while executing wasn't passed
to the calling test.

Now standard and error stream parsing tests may use of all methods
defined in TestExecutionSummary, like:

  assertEquals(2, new ConsoleLauncherWrapper().execute(args).getTestsFoundCount())

instead of former:

  assertTrue("captured out stream as string".contains("2 tests found"))