Apply the 'java' plugin when the `JunitPlatformPlugin` is applied (#709)
Apply the 'java' plugin when the `JunitPlatformPlugin` is applied

Currently, the `JUnitPlatformPlugin` makes assumptions that the
`'java'` plugin has been applied by looking at `sourceSets` and
tasks that are created by it. If the `'java'` plugin is not applied, the
configuration will fail during the `afterEvaluate`.

This change makes it so that the `'java'` plugin is also applied.