Use bottom-up semantics in ReflectionUtils.findMethod()
For clarity of purpose, this commit alters the implementation of the
findMethod() variants in ReflectionUtils so that a bottom-up search
algorithm is used with regard to results returned by the search.

Note, however, that the current algorithm actually traverses the entire
type hierarchy even after having found a matching method. Thus, the
switch from top-down to bottom-up does not have any real effect on the
efficiency of the search algorithm when searching for a single method.

Issue: #731