Only throw JUnitExceptions from AnnotationUtils
Prior to this commit, AnnotationUtils broke a core principle of JUnit
by throwing an AssertionError instead of a JUnitException.

Even though the code block in question should technically never throw
an exception — since it would mean the Java compiler’s support for
repeatable annotations was broken — this commit nonetheless changes the
code to thrown a JUnitException if such a situation occurs.