Introduce support for repeated tests
This commit introduces support for repeating a test a specified number
of times via a new @RepeatedTest annotation.

@RepeatedTest is a @TestTemplate that is supported by a new
RepeatedTestExtension which is automatically registered by default
within the JUnit Jupiter test engine.

In addition to specifying the number of repetitions, a custom display
name can be configured for each repetition via @RepeatedTest.
Furthermore, the display name can be a pattern composed of static text
and dynamic placeholders. The following placeholders are currently
supported.

 - {displayName}: display name of the @RepeatedTest method

 - {currentRepetition}: the current repetition count

 - {totalRepetitions}: the total number of repetitions

The default display name for a given repetition is generated based on
the following pattern: : "{displayName} :: repetition
{currentRepetition} of {totalRepetitions}"

In order to retrieve information about the current repetition and the
total number of repetitions programmatically, a developer can choose to
have an instance of RepetitionInfo injected into a @RepeatedTest,
@BeforeEach, or @AfterEach method.

Issue: #214