Revise validation rules for registered TestTemplateInvocationContextProviders
Prior to this commit, an exception was thrown if there were no
TestTemplateInvocationContextProviders registered when attempting to
execute a @TestTemplate method. An exception was also thrown if none
of the registered providers supports the @TestTemplate method.

It turns out, however, that the first validation rule does not make
sense since we do not yet know if any such provider actually supports
the @TestTemplate method.

This commit addresses this issue by removing the first validation rule,
since the intention behind the first validation rule is covered by the
second validation rule anyway.

Issue: #14