Prohibit display names from being blank
Prior to this commit, the "not blank" precondition for display names
was not consistently enforced.

This commit addresses this issue by documenting that custom displays
names supplied via @DisplayName, @RepeatedTest, and @ParameterizedTest
are never permitted to be blank. Furthermore, the respective handling
of each of these annotations has been updated to log a warning if a
blank display name (or display name pattern) has been supplied by the
user.

At this point in time, it is unfortunately not possible to enforce such
preconditions with an explicit check using Preconditions.isNotBlank()
since we do not have a proper mechanism for handling such configuration
exceptions during the TestEngine discovery phase.

Issue: #743