Ensure generated diagrams are included in User Guide PDF
Prior to this commit, diagrams created using asciidoctor-diagram were
included in the HTML version of the User Guide but not in the PDF
version. This is due to a limitation within the Asciidoctor ecosystem.
Namely, the PDF backend for Asciidoctor does not know about the fact
that asciidoctor-diagram has generated images in a target directory.

This commit addresses this shortcoming by introducing the following
work-around inspired by a tip from @mojavelinux.

- We treat the target directory where asciidoctor-diagram generates
  files as an “image catalog” for both static images and generated
  images.
- We create a custom Copy task that copies all static images to the
  catalog directory.
- The asciidoctor task is configured to depend on the new copy task.
- The ‘imagesdir' and ‘imagesoutdir' attributes of the asciidoctor task
  are both configured to point to the catalog directory.

Issue: #770